move context to separate file
